homeworlds: update stable url, deprecate
The existing `stable` URL for `homeworlds` is a Git revision with no
`tag`. We have some other formulae in a similar situation but they
use a commit tarball instead of checking out the Git repository
(e.g., `luajit`), so this updates the formula to follow the same
pattern.

This also deprecates the formula, as the `stable` URL does not meet
our current requirements.